fix: Update SGD file names to lowercase
