+++ driver/dm_sgd.cpp
@@ -25,15 +25,15 @@
  *******************************************************************************/
 
 #include "registry_driver_maker.hpp"
-#include "SGD_driver.hpp"
+#include "sgd_driver.hpp"
 
 static Driver* makeDriver(const std::string& base_arg)
 {
-    if(base_arg == "SGD")
+    if(base_arg == "sgd")
         return new SGDDriver<float, float>();
-    if(base_arg == "SGDfp16")
+    if(base_arg == "sgdfp16")
         return new SGDDriver<float16, float>();
-    if(base_arg == "SGDbfp16")
+    if(base_arg == "sgdbfp16")
         return new SGDDriver<bfloat16, float>();
     return nullptr;
 }
