+++ src/include/miopen/sgd/problem_description.hpp
@@ -26,8 +26,6 @@
 
 #pragma once
 
-#include "miopen/errors.hpp"
-#include "miopen/miopen.h"
 #include <miopen/activ.hpp>
 #include <miopen/problem_description_base.hpp>
 #include <miopen/tensor.hpp>
